requirements.txt
Django==5.2.7
asgiref==3.10.0
sqlparse==0.5.3
python-dotenv==1.0.1
Pillow==10.4.0
google-cloud-vision==3.7.4
opencv-python-headless==4.10.0.84
numpy>=1.26,<3

